requirements.txt
Django==5.0
djangorestframework==3.14.0
djangorestframework-simplejwt==5.3.0
psycopg2-binary==2.9.9
python-dotenv==1.0.0
django-cors-headers==4.3.0
django-filter==23.5
django-debug-toolbar==4.2.0
django-extensions==3.2.3
django-cleanup==8.0.0
Pillow==10.2.0
django-htmx==1.19.0
django-tailwind==3.8.0
django-allauth==0.58.2
celery==5.3.6
redis==5.0.1
django-redis==5.3.0